Put load and load_from_string back in sfc module
